requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-environ==0.4.5
environ==1.0
idna==2.10
mysqlclient==2.0.3
phonenumbers==8.12.20
prayertime==2.4
python-telegram-bot==13.3
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
